+++ apps/restconf/restconf_main_native.c
@@ -679,7 +679,7 @@ restconf_clixon_backend(clicon_handle h,
     if (clicon_rpc_get_config(h, pw->pw_name, "running", "/restconf", nsc, NULL, &xconfig) < 0)
         goto done;
     if ((xerr = xpath_first(xconfig, NULL, "/rpc-error")) != NULL){
-        clixon_netconf_error(xerr, "Get backend restconf config", NULL);
+        clixon_netconf_error(h, xerr, "Get backend restconf config", NULL);
         goto done;
     }
     /* Extract restconf configuration */
@@ -1017,7 +1017,7 @@ restconf_clixon_init(clicon_handle h,
         if ((ret = clixon_xml_parse_string(inline_config, YB_MODULE, yspec, &xrestconf, &xerr)) < 0)
             goto done;
         if (ret == 0){
-            clixon_netconf_error(xerr, "Inline restconf config", NULL);
+            clixon_netconf_error(h, xerr, "Inline restconf config", NULL);
             goto done;
         }
         /* Replace parent w first child */