+++ util/clixon_util_xpath.c
@@ -326,7 +326,7 @@ main(int    argc,
                 clicon_err(OE_XML, errno, "cbuf_new");
                 goto done;
             }
-            if (netconf_err2cb(xerr, cbret) < 0)
+            if (netconf_err2cb(h, xerr, cbret) < 0)
                 goto done;
             fprintf(stderr, "xml validation error: %s\n", cbuf_get(cbret));
             goto done;
@@ -349,7 +349,7 @@ main(int    argc,
                 clicon_err(OE_XML, errno, "cbuf_new");
                 goto done;
             }
-            if (netconf_err2cb(xerr, cbret) < 0)
+            if (netconf_err2cb(h, xerr, cbret) < 0)
                 goto done;
             fprintf(stderr, "xml validation error: %s\n", cbuf_get(cbret));
             goto done;